poc/pyproject.toml
[project]
name = "taipy-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geopandas>=1.0.1",
    "plotly>=6.0.1",
    "polars>=1.27.1",
    "taipy>=4.0.3",
]
